data-ingestion-service/requirements.txt
uvicorn==0.14.0
fastapi==0.68.0
requests==2.26.0
scikit-learn==0.24.2
wheel==0.36.2
urllib3==1.26.7
setuptools==57.0.0
requests==2.26.0
pymongo==3.12.0
pip==21.1.2
newsapi-python==0.2.6
kafka-python==2.0.2
idna==3.2
charset-normalizer==2.0.6
certifi==2021.5.30